+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Changes
 
+### Next release
+ - Fix tests [#26](https://github.com/okfn/okfn-collaborative-docs/pull/26)
+ - Some help to resync with upstream [#27](https://github.com/okfn/okfn-collaborative-docs/pull/27)
+ - Init and rebase commands [#28](https://github.com/okfn/okfn-collaborative-docs/pull/28)
+
 ### 0.2.1
 2023-01-03 [#24](https://github.com/okfn/okfn-collaborative-docs/pull/24)
  - Tests added